+++ testing/test.py
@@ -161,6 +161,9 @@ cases = [
   Case('split("1.2.3", 1)', UNE_RT_ERROR, UNE_ET_TYPE, []),
   Case('split("1.2.3", [1])', UNE_RT_ERROR, UNE_ET_TYPE, []),
   
+  Case('eval(0)', UNE_RT_ERROR, UNE_ET_TYPE, []),
+  Case('eval("23*2")', UNE_RT_INT, '46', []),
+  
   # main
   Case('unknown', UNE_RT_ERROR, UNE_ET_FILE_NOT_FOUND, [ATTR_DIRECT_ARG]),
   Case('', UNE_RT_ERROR, UNE_ERROR_INPUT, [ATTR_DIRECT_ARG]),